% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes.R
\name{buscar_marmitas_veganas}
\alias{buscar_marmitas_veganas}
\title{Buscar marmitas veganas do cardapio da Beleaf}
\usage{
buscar_marmitas_veganas()
}
\value{
Uma tibble com 5 colunas e 30 linhas
}
\description{
Faz a raspagem dos dados do site da marca Beleaf,
e busca uma tabela contendo informacoes sobre o
cardapio das marmitas veganas.
}
\examples{
buscar_marmitas_veganas()
}
