% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Marmitas}
\alias{Marmitas}
\title{Card\u00e1pio de Marmitas veganas}
\format{
Uma base de dados com 6 colunas
\describe{
\item{$ item          <chr>}{Indice da das marmitas}
\item{$ nome          <chr>}{Nome da marmita}
\item{$ descricao     <chr>}{Descri\u00e7\u00e3o do card\u00e1pio}
\item{$ preco         <dbl>}{Valor da marmita}
\item{$ tipo_refeicao <chr>}{Tipo da marmita: "Sopa", "Prato", "Doce"}
\item{$ url           <chr>}{url para busca da marmita}

...
}
}
\source{
\url{https://www.beleaf.com.br/cardapio-marmitas-veganas}
}
\usage{
Marmitas
}
\description{
Site com card\u00e1pio de marmitas veganas
}
\keyword{datasets}
