% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{marmitas_veganas}
\alias{marmitas_veganas}
\title{Marmitas veganas}
\format{
\describe{
\item{item}{quantidade de itens}
\item{nome}{nome da refeicao}
\item{descricao}{descricao da refeicao}
\item{preco}{preco da refeicao}
\item{tipo_refeicao}{tipo da refeicao}
\item{url}{link da refeicao}
}
}
\source{
\url{https://www.beleaf.com.br/cardapio-marmitas-veganas}
}
\usage{
marmitas_veganas
}
\description{
Essa base possui cardapio de marmitas veganas.
}
\keyword{datasets}
