% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortear_combo_marmitas.R
\name{sortear_combo_marmitas}
\alias{sortear_combo_marmitas}
\title{Sortear combo marmitas}
\usage{
sortear_combo_marmitas(n_marmitas = 12)
}
\arguments{
\item{n_marmitas}{combos de 12, 21 ou 30 marmitas}
}
\value{
uma tibble
}
\description{
Essa funcao ajuda escolher quais marmitas veganas colocar no combo.
}
\examples{
sortear_combo_marmitas()
}
