% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes.R
\name{sortear_marmita}
\alias{sortear_marmita}
\title{Função para sortear uma marmita, usando a base de marmitas}
\usage{
sortear_marmita(tipo = c("Sopa", "Prato", "Doce"), preco_maximo = 50)
}
\arguments{
\item{preco_maximo}{}
}
\value{
Uma tibble com uma marmita para comprar!
}
\description{
Função para sortear uma marmita, usando a base de marmitas
}
\examples{
sortear_marmita()


sortear_marmita(preco_maximo = 22)


sortear_marmita(tipo = "Sopa")


sortear_marmita(tipo = "Prato")


sortear_marmita(tipo = "Doce")

}
