% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscar_marmitas_veganas.R
\name{buscar_marmitas_veganas}
\alias{buscar_marmitas_veganas}
\title{Buscar marmitas veganas}
\usage{
buscar_marmitas_veganas()
}
\value{
uma tibble
}
\description{
Permite obter a base de marmitas veganas.
}
\examples{
buscar_marmitas_veganas()
}
