% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortear_marmita.R
\name{sortear_marmita}
\alias{sortear_marmita}
\title{Sortear marmitas}
\usage{
sortear_marmita(tipo = c("Sopa", "Prato", "Doce"), preco_maximo = 50)
}
\arguments{
\item{tipo}{tipo de refeicao}

\item{preco_maximo}{preco da refeicao}
}
\value{
uma tibble
}
\description{
Funcao para sortear uma marmita vegana.
}
\examples{
sortear_marmita()
}
